refactored SearchArtists page, created ArtistsList component and Loading component
