+++ src/components/ArtistsList/ArtistsList.module.css
@@ -1,0 +1,26 @@
+.artistContainer {
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: space-around;
+  width: 100%;
+  margin-top: 10px;
+  overflow-x: auto;
+  transition: ease-in-out 0.4s;
+}
+
+.artistItem {
+  text-align: center;
+  padding: 10px;
+}
+
+.photoArtist {
+  width: 200px;
+  height: 200px;
+  border-radius: 25px;
+}
+
+.nameSection {
+  color: #555;
+  padding: 5px;
+  margin-top: 5px;
+}
