+++ src/components/Loading/Loading.module.css
@@ -1,0 +1,5 @@
+.loading {
+  color: rgb(101, 90, 163);
+  margin-top: 15px;
+  padding-left: 10px;
+}
