+++ src/views/SearchArtists/SearchArtists.module.css
@@ -26,47 +26,20 @@
   font-size: 16px;
 }
 
-.loading {
-  color: rgb(101, 90, 163);
-  margin-top: 15px;
-  padding-left: 10px;
-}
-
 .errorMessage {
   color: red;
   display: flex;
   align-items: center;
 }
 
-.artistContainer {
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: space-around;
-  width: 100%;
-  margin-top: 10px;
-  overflow-x: auto;
-  transition: ease-in-out 0.4s;
+@media (max-width: 425px) {
+  .searchInput {
+    width: 100%;
+  }
 }
 
-.artistItem {
-  text-align: center;
-  padding: 10px;
-}
-
-.photoContainer {
-}
-
-.photoArtist {
-  width: 200px;
-  height: 200px;
-  border-radius: 25px;
-}
-
-.nameSection {
-  color: #555;
-  padding: 5px;
-  margin-top: 5px;
-}
-
-.artistName {
+@media (min-width: 426px) and (max-width: 768px) {
+  .searchInput {
+    width: 90%;
+  }
 }
